week-2/gemini-chatbot/requirements.txt
fastapi
uvicorn>=0.29
python-multipart
jinja2
google-generativeai>=0.3.0
python-dotenv
python-fasthtml